tests: Remove test/emit_impl.h
